+++ test/recon/BeforeAfter.sol
@@ -9,12 +9,12 @@ import {Governance} from "src/Governance.sol";
 
 abstract contract BeforeAfter is Setup, Asserts {
     struct Vars {
-        uint16 epoch;
+        uint256 epoch;
         mapping(address => IGovernance.InitiativeStatus) initiativeStatus;
         // initiative => user => epoch => claimed
-        mapping(address => mapping(address => mapping(uint16 => bool))) claimedBribeForInitiativeAtEpoch;
-        mapping(address user => uint128 lqtyBalance) userLqtyBalance;
-        mapping(address user => uint128 lusdBalance) userLusdBalance;
+        mapping(address => mapping(address => mapping(uint256 => bool))) claimedBribeForInitiativeAtEpoch;
+        mapping(address user => uint256 lqtyBalance) userLqtyBalance;
+        mapping(address user => uint256 lusdBalance) userLusdBalance;
     }
 
     Vars internal _before;
@@ -27,7 +27,7 @@ abstract contract BeforeAfter is Setup, Asserts {
     }
 
     function __before() internal {
-        uint16 currentEpoch = governance.epoch();
+        uint256 currentEpoch = governance.epoch();
         _before.epoch = currentEpoch;
         for (uint8 i; i < deployedInitiatives.length; i++) {
             address initiative = deployedInitiatives[i];
@@ -38,13 +38,13 @@ abstract contract BeforeAfter is Setup, Asserts {
         }
 
         for (uint8 j; j < users.length; j++) {
-            _before.userLqtyBalance[users[j]] = uint128(lqty.balanceOf(user));
-            _before.userLusdBalance[users[j]] = uint128(lusd.balanceOf(user));
+            _before.userLqtyBalance[users[j]] = uint256(lqty.balanceOf(user));
+            _before.userLusdBalance[users[j]] = uint256(lusd.balanceOf(user));
         }
     }
 
     function __after() internal {
-        uint16 currentEpoch = governance.epoch();
+        uint256 currentEpoch = governance.epoch();
         _after.epoch = currentEpoch;
         for (uint8 i; i < deployedInitiatives.length; i++) {
             address initiative = deployedInitiatives[i];
@@ -55,8 +55,8 @@ abstract contract BeforeAfter is Setup, Asserts {
         }
 
         for (uint8 j; j < users.length; j++) {
-            _after.userLqtyBalance[users[j]] = uint128(lqty.balanceOf(user));
-            _after.userLusdBalance[users[j]] = uint128(lusd.balanceOf(user));
+            _after.userLqtyBalance[users[j]] = uint256(lqty.balanceOf(user));
+            _after.userLusdBalance[users[j]] = uint256(lusd.balanceOf(user));
         }
     }
 }
